Improve ros bag recording (#4)
- `record_rosbag.sh`
 - `max-cache-size` is significantly increased to avoid loosing data
     while writing
  -  rosbags are splitted every 10 GiB 
  - use `mcap_cfg.yaml` for configuring storage
- Data is not compressed to avoid overheads when writing (`compression:
"None"`)
- More topics are added for recording, this includes dataspeed DBW
report topics
- `dev.sh` allows `-p` arg to specity a docker volume where rosbags are
going to be saved locally
- The script will try to use an in-vehicle only directory and will exit
if not found
- Set Cyclone DDS as ROS RMW and configure it for high msg throughput
- Foxglove bridge now runs with `num_threads:=0`

+++ dev.sh
@@ -3,21 +3,59 @@
 # Build docker dev stage and add local code for live development
 # ----------------------------------------------------------------
 
+# Default in-vehicle rosbags directory
+ROSBAGS_DIR=/recorded_datasets/edinburgh
+
+# Function to print usage
+usage() {
+    echo "Usage: dev.sh [-p|--path <absolute path to store rosbags>]"
+    exit 1
+}
+
+# Parse command-line options
+while [[ "$#" -gt 0 ]]; do
+    case $1 in
+            # Option to specify path
+        -p|--path)
+            if [ -n "$2" ]; then
+                ROSBAGS_DIR="$2"
+                shift
+            else
+                echo "Error: Argument for $1 is missing."
+                usage
+            fi
+            ;;
+        -h|--help)
+            usage
+            ;;
+        *)
+            echo "Unknown option: $1"
+            usage
+            ;;
+    esac
+    shift
+done
+
+# Verify ROSBAGS_DIR exists
+if [ ! -d "$ROSBAGS_DIR" ]; then
+    echo "$ROSBAGS_DIR does not exist! Please provide a valid path to store rosbags"
+    exit 1
+fi
+
+
 # Build docker image up to dev stage
 DOCKER_BUILDKIT=1 docker build \
-    -t av_tools_humble:latest \
+    -t av_tools:latest-dev \
     -f Dockerfile --target dev .
 
 # Get the absolute path of the script
 SCRIPT_DIR=$(dirname "$(readlink -f "${BASH_SOURCE[0]}")")
 
-# Create a dir to store rosbags
-mkdir -p rosbags
-
 # Run docker image with local code volumes for development
 docker run -it --rm --net host --privileged \
     -v /dev/shm:/dev/shm \
-    -v $SCRIPT_DIR/rosbags:/opt/ros_ws/rosbags \
+    -v $ROSBAGS_DIR:/opt/ros_ws/rosbags \
     -v $SCRIPT_DIR/scripts/container_tools:/opt/ros_ws/container_tools \
     -v $SCRIPT_DIR/config:/opt/ros_ws/config \
-    av_tools_humble:latest
+    -v /etc/localtime:/etc/localtime:ro \
+    av_tools:latest-dev
